+++ rdf-abac-core/src/test/java/io/telicent/jena/abac/TS_ABAC_Core.java
@@ -43,9 +43,10 @@ import org.junit.platform.suite.api.Suite;
     , TestLabelsMem.class
     , TestLabelsMemNoPatterns.class
     , TestAssemblerABAC.class
-
     , TestLabelsStoreMem.class
     , TestTransactionalMemory.class
+    , TestAllNamedGraphs.class
+    , TestAllNamedGraphsTdb2.class
     , TestAE.class
     , TestABAC.class
     , TestCtxABAC.class